backend/requirements.txt
fastapi==0.118.0
loguru==0.7.3
pydantic==2.11.9
python-dotenv==1.1.1
psycopg2==2.9.9
redis==3.5.3
Requests==2.32.5
SQLAlchemy==2.0.23
starlette==0.48.0
uvicorn==0.37.0
PyYAML==6.0.3